+++ src/utils/loader.ts
@@ -43,6 +43,7 @@ import StatsCommand from '../commands/stats';
 import ExecCommand from '../commands/exec';
 import ProfileCommand from '../commands/profile';
 import AutoRoleCommand from '../commands/autorole';
+import GreetingCommand from '../commands/greeting';
 
 export const loadEvents = (bot: Bot): void => {
     bot.events.push(new ReadyEvent, new MessageEvent, new GuildCreateEvent,
@@ -64,7 +65,7 @@ export const loadCommands = (bot: Bot): void => {
         new QueueCommand, new BanCommand, new ServerInfoCommand,
         new PermissionsCommand, new KickCommand, new StopCommand,
         new PrefixCommand, new StatsCommand, new ExecCommand, new ProfileCommand,
-        new AutoRoleCommand);
+        new AutoRoleCommand, new GreetingCommand);
 
     console.log(blue(`Loaded ${bot.commands.length} commands`));
 }